pyproject.toml
[project]
name = "yiyin"
version = "0.1.0"
description = "yiyin"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi,httpx,websockets]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "pillow>=12.1.1",
]

[tool.nonebot]
plugin_dirs = ["yiyin"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
